backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
supabase>=2.4
requests>=2.31
openai>=1.12
python-dotenv>=1.0